Add dependency for floating IP
The hot template for N-S with FIP, works fine in most cases. However, when
setting the router properties to distributed: true, attaching a floating ip
almost always certainly fails with the message that floating ip couldn't be
associated as the external subnet is not reachable from the private subnet.
The BZ is https://bugs.launchpad.net/heat/+bug/1626619
Making the floatingip depend on the router_interface_2 will ensure we get
successfull stack deployments and deletes.

Change-Id: I20b0786470a34d121b6c431d9e01d7fd8d59b363