Use try-except block for parameters for heat stack
This commit adds a try-except block when populating parameters for
creating the heat stack. For example, if a user forgets to create an
external network, we throw an exception in
shaker.openstack.clients.neutron.choose_external_net, however we continue
even though connection to openstack would fail in server.py to accomodate
for scenarios that do not need OpenStack connection. We need to fail and
exit before creating heat stack if one of the required parameters is missing.

Change-Id: Ic5a10a83f51dc7f0c17d31eeca5a94ade8ee1409