Enhance Iperf executor to accept "unmapped" args
Issue: The current iperf executor is setup nicely to handle the
majority, let's say 90%, of the test cases a user might want to run.
However, it still leaves a bit to be desired for the remaining 10%. For
example we use a custom qcow2 image for our tests  and the verison of
iperf in this image (and the latest iperf which shaker doesn't have) you
can specify --sctp as an argument to run this type of traffic. This is
not possible with the current executor, but most users may not want or
care about this ability.

This commit proposes to extend the current executor to accept any
argument as part of an "advanced" mode, rather than mapping every
argument that may or may not used or exists only in certain versions.
This allows different version of iperf (e.g. --sctp capable) of
executing a supported argument without having to directly add support
for this new argument into the executor

I know that any argument could be passed to iperf using the script
executor, but this has the downside of not providing the nice graphs.
I believe this commit is a good compromise between general
useability (pre-mapping arguments) and granular control (using any
agrument that is not mapped). This will also be useful for other
arguments that may not be widely used, or are newly introduced, such
as --zerocopy or --omit along with other sctp related argument like
--xbind or --nstreams.

This would be for advanced users familiar with iperf and would not
impact current tests or general use. The only "safety net" provided is
not allowing mutually exclusive arguments with time to be specified with
the time argument. Since time is currently added by default, even if we
were to create as mapping for say --blockcount(-k), we would have to
handle removing the time argument for a correct run. Arguments that
don't mix will lead to an iperf error, but this error is correctly
logged and handled. Overriding time if a mutually exclusive argument is
used seems like a fair way to handle the situation while allow other
possible argument combinations.

Change-Id: I072278d0ab3912aac375605dc36987a7f8c20461