Make {{ unique }} available in env files
Issue: For certain tests we need to create long strings that use the
stack's name. Currently we have no way to inject the stack name into the
env file to build these necessary strings as parameters to the
templates.

This commit allows users to reference the dynamic stack name in the env
file the same way it's referenced in the Heat template by using {{
unique }}

Change-Id: I817189ac6353e34fe7cad84bbaf484c30b6b97b2